Move OnboardUIState.kt to com.tpp.theperiodpurse.ui.state
