+++ app/src/main/java/com/tpp/theperiodpurse/ui/onboarding/DateRangePicker.kt
@@ -17,7 +17,7 @@ import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.zIndex
-import com.tpp.theperiodpurse.data.OnboardUIState
+import com.tpp.theperiodpurse.ui.state.OnboardUIState
 import com.tpp.theperiodpurse.ui.viewmodel.OnboardViewModel
 import java.time.Instant
 import java.time.LocalDate
