+++ app/src/main/java/com/tpp/theperiodpurse/ui/onboarding/QuestionThreeScreen.kt
@@ -28,7 +28,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavHostController
 import com.tpp.theperiodpurse.OnboardingScreen
-import com.tpp.theperiodpurse.data.OnboardUIState
+import com.tpp.theperiodpurse.ui.state.OnboardUIState
 
 @RequiresApi(Build.VERSION_CODES.O)
 @Composable
