+++ app/src/main/java/com/tpp/theperiodpurse/ui/setting/ResetDatabase.kt
@@ -11,7 +11,7 @@ import androidx.compose.runtime.livedata.observeAsState
 import androidx.navigation.NavHostController
 import com.tpp.theperiodpurse.ui.state.AppUiState
 import com.tpp.theperiodpurse.OnboardingScreen
-import com.tpp.theperiodpurse.data.OnboardUIState
+import com.tpp.theperiodpurse.ui.state.OnboardUIState
 import com.tpp.theperiodpurse.ui.state.CalendarUIState
 import com.tpp.theperiodpurse.ui.onboarding.LoadingScreen
 import com.tpp.theperiodpurse.ui.viewmodel.OnboardViewModel
