+++ app/src/main/java/com/tpp/theperiodpurse/ui/viewmodel/OnboardViewModel.kt
@@ -18,7 +18,7 @@ import com.tpp.theperiodpurse.data.entity.User
 import com.tpp.theperiodpurse.data.model.Symptom
 import com.tpp.theperiodpurse.data.repository.DateRepository
 import com.tpp.theperiodpurse.data.repository.UserRepository
-import com.tpp.theperiodpurse.data.OnboardUIState
+import com.tpp.theperiodpurse.ui.state.OnboardUIState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.*
